[Refactor] reorganize logical optimizer into optimizer.ir package
- move ops, dataframe ops, and numeric ops into the new stratum.optimizer.ir namespace
- update runtime, benchmarks, and tests to import from the optimizer package
- refresh README and benchmark paths to reflect the new optimizer layout
