+++ stratum/optimizer/__init__.py
