+++ stratum/optimizer/_cse.py
