+++ stratum/optimizer/_op_comparison.py
