+++ stratum/optimizer/rewrite_ideas.md
