+++ stratum/runtime/_scheduler.py
@@ -3,9 +3,9 @@ from sklearn.metrics import mean_squared_error
 from sklearn.model_selection import train_test_split, check_cv
 from sklearn.metrics._scorer import _Scorer, get_scorer
 from skrub._data_ops._data_ops import EvalMode
-from stratum.logical_optimizer._dataframe_ops import SplitOp
-from stratum.logical_optimizer._op_utils import topological_iterator
-from stratum.logical_optimizer._ops import ImplOp, Op
+from stratum.optimizer.ir._dataframe_ops import SplitOp
+from stratum.optimizer._op_utils import topological_iterator
+from stratum.optimizer.ir._ops import ImplOp, Op
 import polars as pl
 
 import logging
