+++ stratum/tests/logical_optimizer/algebraic_rewrites/test_numeric.py
@@ -1,10 +1,10 @@
 import unittest
 import stratum as skrub
 import numpy as np
-from stratum.logical_optimizer._optimize import  optimize
-from stratum.logical_optimizer._optimize import OptConfig
-from stratum.logical_optimizer._algebraic_rewrites import AlgebraicRewritesConfig
-from stratum.logical_optimizer._op_utils import topological_iterator
+from stratum.optimizer._optimize import  optimize
+from stratum.optimizer._optimize import OptConfig
+from stratum.optimizer._algebraic_rewrites import AlgebraicRewritesConfig
+from stratum.optimizer._op_utils import topological_iterator
 
 class TestCSE(unittest.TestCase):
