+++ stratum/tests/logical_optimizer/test_op_utils.py
@@ -1,8 +1,8 @@
 #from curses import flash
 import unittest
 import stratum as skrub
-from stratum.logical_optimizer._optimize import optimize as optimize_, OptConfig, choice_unrolling
-from stratum.logical_optimizer._op_utils import show_graph, clone_sub_dag, topological_iterator, FLAGS
+from stratum.optimizer._optimize import optimize as optimize_, OptConfig, choice_unrolling
+from stratum.optimizer._op_utils import show_graph, clone_sub_dag, topological_iterator, FLAGS
 from stratum._config import config
 graph = False
